requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
beautifulsoup4==4.9.1     # via bs4
bs4==0.0.1                # via data_finder (setup.py)
certifi==2020.6.20        # via requests
chardet==3.0.4            # via requests
click==7.1.2              # via data_finder (setup.py)
idna==2.10                # via requests
pillow==7.2.0             # via data_finder (setup.py)
pymupdf==1.17.7           # via data_finder (setup.py)
python-magic==0.4.18      # via data_finder (setup.py)
requests==2.24.0          # via data_finder (setup.py)
soupsieve==2.0.1          # via beautifulsoup4
sqlalchemy==1.3.19        # via data_finder (setup.py)
urllib3==1.25.10          # via requests
